Fixed opening notif

+++ src/main/dotnet/WinWiFi/JavaInterop/Interop.cs
@@ -11,7 +11,7 @@ namespace JavaInterop
             wifi = new Wifi();
             string rawInput = "";
             int input;
-            Console.WriteLine(ResponseToJava.SESSION_OPENED);
+            Console.WriteLine((int)ResponseToJava.SESSION_OPENED);
             do
             {
                 rawInput = Console.ReadLine();
